+++ connector/src/test/java/it/eng/connector/integration/applicationproperties/ApplicationPropertyIntegrationTest.java
@@ -1,6 +1,5 @@
 package it.eng.connector.integration.applicationproperties;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -8,7 +7,9 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import org.junit.jupiter.api.Test;
@@ -21,7 +22,9 @@ import com.fasterxml.jackson.core.type.TypeReference;
 
 import it.eng.connector.integration.BaseIntegrationTest;
 import it.eng.connector.util.TestUtil;
+import it.eng.tools.controller.ApiEndpoints;
 import it.eng.tools.model.ApplicationProperty;
+import it.eng.tools.property.ApplicationPropertyKeys;
 import it.eng.tools.repository.ApplicationPropertiesRepository;
 import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.serializer.ToolsSerializer;
@@ -36,9 +39,9 @@ public class ApplicationPropertyIntegrationTest extends BaseIntegrationTest {
 	@Test
 	@WithUserDetails(TestUtil.ADMIN_USER)
 	public void getPropertiesSuccessfulTest() throws Exception {
-		final ResultActions result =
+		ResultActions result =
 				mockMvc.perform(
-						get("/api/v1/properties/")
+						get(ApiEndpoints.PROPERTIES_V1 + "/")
 						.contentType(MediaType.APPLICATION_JSON_VALUE)
 						.accept(MediaType.APPLICATION_JSON_VALUE));
 
@@ -51,32 +54,21 @@ public class ApplicationPropertyIntegrationTest extends BaseIntegrationTest {
 		  
 		assertNotNull(apiResp.getData());
 		assertTrue(apiResp.getData().size() > 2);
-	}
-
-	@Test
-	@WithUserDetails(TestUtil.ADMIN_USER)
-	public void getPropertySuccessfulTest() throws Exception {
-		ApplicationProperty property = ApplicationProperty.Builder.newInstance()
-				.key(TEST_KEY)
-				.value("abc")
-				.build();
-		repository.save(property);
-
-		final ResultActions result =
-				mockMvc.perform(
-						get("/api/v1/properties/{key}", this.TEST_KEY )
-						.contentType(MediaType.APPLICATION_JSON_VALUE));
-
-		result.andExpect(status().isOk())
-		.andExpect(content().contentType(MediaType.APPLICATION_JSON));
 		
-		String json = result.andReturn().getResponse().getContentAsString();
-		TypeReference<GenericApiResponse<ApplicationProperty>> typeRef = new TypeReference<GenericApiResponse<ApplicationProperty>>() {};
-		GenericApiResponse<ApplicationProperty> apiResp =  ToolsSerializer.deserializePlain(json, typeRef);
+		result =
+				mockMvc.perform(
+						get(ApiEndpoints.PROPERTIES_V1 + "/?key_prefix=" + ApplicationPropertyKeys.DAPS_PREFIX)
+						.contentType(MediaType.APPLICATION_JSON_VALUE)
+						.accept(MediaType.APPLICATION_JSON_VALUE));
+		result.andExpect(status().isOk())
+			.andExpect(content().contentType(MediaType.APPLICATION_JSON));
+		
+		json = result.andReturn().getResponse().getContentAsString();
+		apiResp =  ToolsSerializer.deserializePlain(json, typeRef);
 		  
 		assertNotNull(apiResp.getData());
-		assertEquals(apiResp.getData().getKey(), TEST_KEY);
-		repository.delete(property);
+		Optional<ApplicationProperty> shouldBeEmpty = apiResp.getData().stream().filter(prop -> !prop.getKey().contains(ApplicationPropertyKeys.DAPS_PREFIX)).findAny();
+		assertTrue(shouldBeEmpty.isEmpty());
 	}
 
 	@Test
@@ -95,7 +87,7 @@ public class ApplicationPropertyIntegrationTest extends BaseIntegrationTest {
 				.value(randomValue)
 				.build();
 
-		String body = ToolsSerializer.serializeProtocolJsonNode(changedProperty).toString();
+		String body = ToolsSerializer.serializePlain(Arrays.asList(changedProperty)).toString();
 
 		final ResultActions result =
 				mockMvc.perform(
@@ -108,8 +100,8 @@ public class ApplicationPropertyIntegrationTest extends BaseIntegrationTest {
 			.andExpect(content().contentType(MediaType.APPLICATION_JSON));
 
 		String json = result.andReturn().getResponse().getContentAsString();
-		TypeReference<GenericApiResponse<ApplicationProperty>> typeRef = new TypeReference<GenericApiResponse<ApplicationProperty>>() {};
-		GenericApiResponse<ApplicationProperty> apiResp =  ToolsSerializer.deserializePlain(json, typeRef);
+		TypeReference<GenericApiResponse<List<ApplicationProperty>>> typeRef = new TypeReference<GenericApiResponse<List<ApplicationProperty>>>() {};
+		GenericApiResponse<List<ApplicationProperty>> apiResp = ToolsSerializer.deserializePlain(json, typeRef);
 		  
 		assertNotNull(apiResp.getData());
 		repository.deleteById(changedProperty.getKey());