+++ tools/src/main/java/it/eng/tools/daps/DapsProperties.java
@@ -14,6 +14,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 
 import it.eng.tools.exception.DapsPropertyErrorException;
+import it.eng.tools.property.ApplicationPropertyKeys;
 import it.eng.tools.service.ApplicationPropertiesService;
 import lombok.extern.slf4j.Slf4j;
 
@@ -21,26 +22,14 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class DapsProperties {
 
-	private final String PREFIX = "application.daps.";
-	private final String ENABLED_DAPS_INTERACTION_KEY = PREFIX + "enabledDapsInteraction";
-	private final String EXTENDED_TOKEN_VALIDATION_KEY = PREFIX + "extendedTokenValidation";
-	private final String TOKEN_CACHING_KEY = PREFIX + "tokenCaching";
-	private final String FETCH_TOKEN_ON_STARTUP_KEY = PREFIX + "fetchTokenOnStartup";
-	private final String DAPS_URL_KEY = PREFIX + "dapsUrl";
-	private final String DAPS_JWKS_URL_KEY = PREFIX + "dapsJWKSUrl";
-	private final String DAPS_KEYSTORE_NAME_KEY = PREFIX + "dapsKeyStoreName";
-	private final String DAPS_KEYSTORE_PASSWORD_KEY = PREFIX + "dapsKeyStorePassword";
-	private final String DAPS_KEYSTORE_ALIAS_NAME_KEY = PREFIX + "dapsKeystoreAliasName";
-
 	private final ApplicationPropertiesService service;
 
 	public DapsProperties(ApplicationPropertiesService service) {
 		this.service = service;
 	}
 
-
 	protected boolean isEnabledDapsInteraction() {
-		String enabledDapsInteractionValue = service.get(ENABLED_DAPS_INTERACTION_KEY);
+		String enabledDapsInteractionValue = service.get(ApplicationPropertyKeys.ENABLED_DAPS_INTERACTION_KEY);
 
 		if(enabledDapsInteractionValue != null) {
 			return Boolean.valueOf(enabledDapsInteractionValue);
@@ -49,7 +38,7 @@ public class DapsProperties {
 	}
 
 	protected boolean isExtendedTokenValidation() {
-		String extendedTokenValidationValue = service.get(EXTENDED_TOKEN_VALIDATION_KEY);
+		String extendedTokenValidationValue = service.get(ApplicationPropertyKeys.EXTENDED_TOKEN_VALIDATION_KEY);
 
 		if(extendedTokenValidationValue != null) {
 			return Boolean.valueOf(extendedTokenValidationValue);
@@ -58,7 +47,7 @@ public class DapsProperties {
 	}
 
 	protected boolean isTokenCaching() {
-		String tokenCachingValue = service.get(TOKEN_CACHING_KEY);
+		String tokenCachingValue = service.get(ApplicationPropertyKeys.TOKEN_CACHING_KEY);
 
 		if(tokenCachingValue != null) {
 			return Boolean.valueOf(tokenCachingValue);
@@ -67,7 +56,7 @@ public class DapsProperties {
 	}
 
 	protected boolean isFetchTokenOnStartup() {
-		String fetchTokenOnStartupValue = service.get(FETCH_TOKEN_ON_STARTUP_KEY);
+		String fetchTokenOnStartupValue = service.get(ApplicationPropertyKeys.FETCH_TOKEN_ON_STARTUP_KEY);
 
 		if(fetchTokenOnStartupValue != null) {
 			return Boolean.valueOf(fetchTokenOnStartupValue);
@@ -76,56 +65,56 @@ public class DapsProperties {
 	}
 
 	protected URL getDapsUrl() {
-		String dapsUrlValue = service.get(DAPS_URL_KEY);
+		String dapsUrlValue = service.get(ApplicationPropertyKeys.DAPS_URL_KEY);
 
 		URL dapsUrl = null;
 		try {
 			dapsUrl = new URL(dapsUrlValue);
 		} catch (MalformedURLException e) {
-			throw new DapsPropertyErrorException(DAPS_URL_KEY + " isn't defined!!!");
+			throw new DapsPropertyErrorException(ApplicationPropertyKeys.DAPS_URL_KEY + " isn't defined!!!");
 		}
 
 		return dapsUrl;
 	}
 
 	protected URL getDapsJWKSUrl() {
-		String dapsJWKSUrlValue = service.get(DAPS_JWKS_URL_KEY);
+		String dapsJWKSUrlValue = service.get(ApplicationPropertyKeys.DAPS_JWKS_URL_KEY);
 
 		URL dapsJWKSUrl = null;
 		try {
 			dapsJWKSUrl = new URL(dapsJWKSUrlValue);
 		} catch (MalformedURLException e) {
-			throw new DapsPropertyErrorException(DAPS_JWKS_URL_KEY + " isn't defined!!!");
+			throw new DapsPropertyErrorException(ApplicationPropertyKeys.DAPS_JWKS_URL_KEY + " isn't defined!!!");
 		}
 
 		return dapsJWKSUrl;
 	}
 
 	protected String getDapsKeyStoreName() {
-		String dapsKeyStoreNameValue = service.get(DAPS_KEYSTORE_NAME_KEY);
+		String dapsKeyStoreNameValue = service.get(ApplicationPropertyKeys.DAPS_KEYSTORE_NAME_KEY);
 
 		if(dapsKeyStoreNameValue == null || dapsKeyStoreNameValue.isEmpty()) {
-			throw new DapsPropertyErrorException(DAPS_KEYSTORE_NAME_KEY + " isn't defined!!!");
+			throw new DapsPropertyErrorException(ApplicationPropertyKeys.DAPS_KEYSTORE_NAME_KEY + " isn't defined!!!");
 		}
 
 		return dapsKeyStoreNameValue;
 	}
 
 	protected String getDapsKeyStorePassword() {
-		String dapsKeyStorePasswordValue = service.get(DAPS_KEYSTORE_PASSWORD_KEY);
+		String dapsKeyStorePasswordValue = service.get(ApplicationPropertyKeys.DAPS_KEYSTORE_PASSWORD_KEY);
 
 		if(dapsKeyStorePasswordValue == null || dapsKeyStorePasswordValue.isEmpty()) {
-			throw new DapsPropertyErrorException(DAPS_KEYSTORE_PASSWORD_KEY + " isn't defined!!!");
+			throw new DapsPropertyErrorException(ApplicationPropertyKeys.DAPS_KEYSTORE_PASSWORD_KEY + " isn't defined!!!");
 		}
 
 		return dapsKeyStorePasswordValue;
 	}
 
 	protected String getDapsKeystoreAliasName() {
-		String dapsKeystoreAliasNameValue = service.get(DAPS_KEYSTORE_ALIAS_NAME_KEY);
+		String dapsKeystoreAliasNameValue = service.get(ApplicationPropertyKeys.DAPS_KEYSTORE_ALIAS_NAME_KEY);
 
 		if(dapsKeystoreAliasNameValue == null || dapsKeystoreAliasNameValue.isEmpty()) {
-			throw new DapsPropertyErrorException(DAPS_KEYSTORE_ALIAS_NAME_KEY + " isn't defined!!!");
+			throw new DapsPropertyErrorException(ApplicationPropertyKeys.DAPS_KEYSTORE_ALIAS_NAME_KEY + " isn't defined!!!");
 		}
 
 		return dapsKeystoreAliasNameValue;